+++ src/services/baseService.test.ts
@@ -61,7 +61,7 @@ describe("Base Service", () => {
     mockFs.restore();
   });
 
-  function createTestService(options?: Serverless.Options) {
+  function createMockService(options?: Serverless.Options) {
     sls = MockFactory.createTestServerless();
     sls.variables["azureCredentials"] = MockFactory.createTestAzureCredentials();
     sls.variables["subscriptionId"] = "ABC123";
@@ -71,7 +71,7 @@ describe("Base Service", () => {
   }
 
   beforeEach(() => {
-    service = createTestService();
+    service = createMockService();
   });
 
   it("Initializes common service properties", () => {
@@ -85,9 +85,9 @@ describe("Base Service", () => {
   });
 
   it("Sets default region and stage values if not defined", () => {
-    const testService = new MockService(sls);
+    const mockService = new MockService(sls);
 
-    expect(testService).not.toBeNull();
+    expect(mockService).not.toBeNull();
     expect(sls.service.provider.region).toEqual("westus");
     expect(sls.service.provider.stage).toEqual("dev");
   });
@@ -97,16 +97,16 @@ describe("Base Service", () => {
       stage: "prod",
       region: "eastus2"
     };
-    const testService = new MockService(sls, cliOptions);
+    const mockService = new MockService(sls, cliOptions);
 
-    expect(testService.getSlsRegion()).toEqual(cliOptions.region);
-    expect(testService.getSlsStage()).toEqual(cliOptions.stage);
+    expect(mockService.getSlsRegion()).toEqual(cliOptions.region);
+    expect(mockService.getSlsStage()).toEqual(cliOptions.stage);
   });
 
   it("Sets default region and stage values if not defined", () => {
-    const testService = new MockService(sls);
+    const mockService = new MockService(sls);
 
-    expect(testService).not.toBeNull();
+    expect(mockService).not.toBeNull();
     expect(sls.service.provider.region).toEqual("westus");
     expect(sls.service.provider.stage).toEqual("dev");
   });
@@ -116,25 +116,25 @@ describe("Base Service", () => {
       stage: "prod",
       region: "eastus2"
     };
-    const testService = new MockService(sls, cliOptions);
+    const mockService = new MockService(sls, cliOptions);
 
-    expect(testService.getSlsRegion()).toEqual(cliOptions.region);
-    expect(testService.getSlsStage()).toEqual(cliOptions.stage);
+    expect(mockService.getSlsRegion()).toEqual(cliOptions.region);
+    expect(mockService.getSlsStage()).toEqual(cliOptions.stage);
   });
 
   it("Generates resource group name from sls yaml config", () => {
-    const testService = new MockService(sls);
-    const resourceGroupName = testService.getSlsResourceGroupName();
+    const mockService = new MockService(sls);
+    const resourceGroupName = mockService.getSlsResourceGroupName();
 
     expect(resourceGroupName).toEqual(sls.service.provider["resourceGroup"]);
   });
 
   it("Generates resource group from convention when NOT defined in sls yaml", () => {
     sls.service.provider["resourceGroup"] = null;
-    const testService = new MockService(sls);
-    const resourceGroupName = testService.getSlsResourceGroupName();
-    const region = testService.getSlsRegion();
-    const stage = testService.getSlsStage();
+    const mockService = new MockService(sls);
+    const resourceGroupName = mockService.getSlsResourceGroupName();
+    const region = mockService.getSlsRegion();
+    const stage = mockService.getSlsStage();
 
     expect(resourceGroupName).toEqual(`sls-${region}-${stage}-${sls.service["service"]}-rg`);
   });
